Separate main from library
